+++ R/modify_table1.R
@@ -33,6 +33,102 @@ charset.px <- function(x, value) {
 }
 
 
+#' @rdname aggregallowed.px
+#' @export
+aggregallowed <- function(x, value) {
+  UseMethod("aggregallowed")
+}
+
+#' @inherit charset.px
+#' @title AGGREGALLOWED
+#' @description `r table_description("AGGREGALLOWED")`
+#' @param value `r table1_param_value("AGGREGALLOWED")`
+#' @export
+aggregallowed.px <- function(x, value) {
+  handle_table1_keyword(x, value, "AGGREGALLOWED")
+}
+
+
+#' @rdname autopen.px
+#' @export
+autopen <- function(x, value) {
+  UseMethod("autopen")
+}
+
+#' @inherit charset.px
+#' @title AUTOPEN
+#' @description `r table_description("AUTOPEN")`
+#' @param value `r table1_param_value("AUTOPEN")`
+#' @export
+autopen.px <- function(x, value) {
+  handle_table1_keyword(x, value, "AUTOPEN")
+}
+
+
+#' @rdname axis_version.px
+#' @export
+axis_version <- function(x, value) {
+  UseMethod("axis_version")
+}
+
+#' @inherit charset.px
+#' @title AXIS-VERSION
+#' @description `r table_description("AXIS-VERSION")`
+#' @param value `r table1_param_value("AXIS-VERSION")`
+#' @export
+axis_version.px <- function(x, value) {
+  handle_table1_keyword(x, value, "AXIS-VERSION")
+}
+
+
+#' @rdname codepage.px
+#' @export
+codepage <- function(x, value) {
+  UseMethod("codepage")
+}
+
+#' @inherit charset.px
+#' @title CODEPAGE
+#' @description `r table_description("CODEPAGE")`
+#' @param value `r table1_param_value("CODEPAGE")`
+#' @export
+codepage.px <- function(x, value) {
+  handle_table1_keyword(x, value, "CODEPAGE")
+}
+
+
+#' @rdname confidential.px
+#' @export
+confidential <- function(x, value) {
+  UseMethod("confidential")
+}
+
+#' @inherit charset.px
+#' @title CONFIDENTIAL
+#' @description `r table_description("CONFIDENTIAL")`
+#' @param value `r table1_param_value("CONFIDENTIAL")`
+#' @export
+confidential.px <- function(x, value) {
+  handle_table1_keyword(x, value, "CONFIDENTIAL")
+}
+
+
+#' @rdname copyright.px
+#' @export
+copyright <- function(x, value) {
+  UseMethod("copyright")
+}
+
+#' @inherit charset.px
+#' @title COPYRIGHT
+#' @description `r table_description("COPYRIGHT")`
+#' @param value `r table1_param_value("COPYRIGHT")`
+#' @export
+copyright.px <- function(x, value) {
+  handle_table1_keyword(x, value, "COPYRIGHT")
+}
+
+
 #' @rdname creation_date.px
 #' @export
 creation_date <- function(x, value) {
@@ -97,6 +193,22 @@ next_update.px <- function(x, value) {
 }
 
 
+#' @rdname showdecimals.px
+#' @export
+showdecimals <- function(x, value) {
+  UseMethod("showdecimals")
+}
+
+#' @inherit charset.px
+#' @title SHOWDECIMALS
+#' @description `r table_description("SHOWDECIMALS")`
+#' @param value `r table1_param_value("SHOWDECIMALS")`
+#' @export
+showdecimals.px <- function(x, value) {
+  handle_table1_keyword(x, value, "SHOWDECIMALS")
+}
+
+
 #' @rdname subject_code.px
 #' @export
 subject_code <- function(x, value) {
@@ -110,4 +222,36 @@ subject_code <- function(x, value) {
 #' @export
 subject_code.px <- function(x, value) {
   handle_table1_keyword(x, value, "SUBJECT-CODE")
+}
+
+
+#' @rdname tableid.px
+#' @export
+tableid <- function(x, value) {
+  UseMethod("tableid")
+}
+
+#' @inherit charset.px
+#' @title TABLEID
+#' @description `r table_description("TABLEID")`
+#' @param value `r table1_param_value("TABLEID")`
+#' @export
+tableid.px <- function(x, value) {
+  handle_table1_keyword(x, value, "TABLEID")
+}
+
+
+#' @rdname update_frequency.px
+#' @export
+update_frequency <- function(x, value) {
+  UseMethod("update_frequency")
+}
+
+#' @inherit charset.px
+#' @title UPDATE-FREQUENCY
+#' @description `r table_description("UPDATE-FREQUENCY")`
+#' @param value `r table1_param_value("UPDATE-FREQUENCY")`
+#' @export
+update_frequency.px <- function(x, value) {
+  handle_table1_keyword(x, value, "UPDATE-FREQUENCY")
 }